add a test to make sure animation data checks out
