--- src/math/operations/resizeOps/index.ts
@@ -1,3 +1,0 @@
-export { expand, snub, dual } from "./expand"
-export { contract } from "./contract"
-export { twist } from "./twist"
